v2.x - move MessageRef to well-known/
