+++ src/index.js
@@ -1,7 +1,6 @@
 import Field from './Field';
 import FieldBuilder from './FieldBuilder';
 import Message from './Message';
-import MessageRef from './MessageRef';
 import Mixin from './Mixin';
 import Schema from './Schema';
 import SchemaCurie from './SchemaCurie';
@@ -15,7 +14,6 @@ export default {
   Field,
   FieldBuilder,
   Message,
-  MessageRef,
   Mixin,
   Schema,
   SchemaCurie,
