+++ src/well-known/index.js
@@ -3,6 +3,7 @@ import DatedSlugIdentifier from './DatedSlugIdentifier';
 import DynamicField from './DynamicField';
 import GeoPoint from './GeoPoint';
 import Identifier from './Identifier';
+import MessageRef from './MessageRef';
 import Microtime from './Microtime';
 import SlugIdentifier from './SlugIdentifier';
 import TimeUuidIdentifier from './TimeUuidIdentifier';
@@ -14,6 +15,7 @@ export default {
   DynamicField,
   GeoPoint,
   Identifier,
+  MessageRef,
   Microtime,
   SlugIdentifier,
   TimeUuidIdentifier,
